AST CHANGES TO DECLARATIONS -- READ FULL DESCRIPTION
This commit makes some MAJOR changes to the AST.

It removes the ability to declare `fn`s on a
non-global scope, and does the same to `struct`s.

The reasoning behind this is because the current
design of the `CgScope` struct in zrc_codegen is
built under the assumption you will only create a
subscope of it, and this is not true once you consider
nested functions. They must exist in the _global_
scope, so that they can access globals. CgScope and
the zrc_typeck Scope do not provide a method of returning
to the global scope and I do not want to implement that
currently.

For this reason, structs and functions will always exist in
the global scope. This makes handling them a separate case
so there is no restoring of scope in codegen or typeck.

This commit also removes the ability to declare globals.
This is quite a difficult removal to make, but it is to
simplify code generation until we are ready to handle
compiling expressions into LLVM ConstExprs.

This will all be undone in the future. It is just to prevent
bugs and simplify the process for now.