make it so only lvalues are callable
This may be seen as a very odd choice, but I cannot see a place
where anything but an lvalue will be called. In addition, there's
a reason: If treated as a place, we return a pointer. If treated
as an expression, we end up *loading* the pointer for identifiers,
and loading a function pointer doesn't make much sense.