refactor: remove data passing from make_test_prelude_closure
Originally, the test prelude system supported "data passing" -- the
return of arbitrary data from your prelude closure that would be passed
downstream. However, this was only used for passing FunctionType
instances of called extern functions so we cound later use
.build_indirect_call to call them (because normal calls use the indirect
method and it generates a slightly different IR). However, it turns out
that LLVM will use .build_call *for us* under the scenes, so there is no
need to use indirect calls in the `expected` IRs. The function types
there can be obtained via a method on module, so we remove the data
passing.