switch `zrc_codegen` to use inkwell/LLVM APIs instead of generating as a string (#82)
This is a squash of... a lot of commits. You can see the reality of each one in PR #82.

* initial cg_expr function for inkwell

* rewrite hopefully all codegen to inkwell

* format

* oops

* make it run successfully

* test action fix

* add llvm to all actions

* fix a couple lints

* fix try_fold lint

* last lints

* add verify pass

* add stub test

* stuff

* some (bad) docs

* r

* r

* docs

* lint fixes

* add first test

* chore: format

* trigger ci

* horrid code but i desperately need a checkpoint

* fix lints and format

* add `other_deref_generates_as_expected`

* add `pointer_indexing_generates_proper_gep`

* add gep test

* comma op test

* add two more!

* add ternary expr test

* let decl test

* test prelude v2

* chore: format

* first two cgb tests

* add a huge while loop test that happens to also test if conditions

* fmt

* refactor: remove data passing from make_test_prelude_closure

Originally, the test prelude system supported "data passing" -- the
return of arbitrary data from your prelude closure that would be passed
downstream. However, this was only used for passing FunctionType
instances of called extern functions so we cound later use
.build_indirect_call to call them (because normal calls use the indirect
method and it generates a slightly different IR). However, it turns out
that LLVM will use .build_call *for us* under the scenes, so there is no
need to use indirect calls in the `expected` IRs. The function types
there can be obtained via a method on module, so we remove the data
passing.

* move if/else tests into their own `conditionals` module

* add test for end block creation

* rename arg

* fix: fix invalid cg on terminating if/else with no trailer

In the case of `if (x) return; else return;` an empty bb was created.

* fix: properly position the builder in `while` loops

This caused the rest of the exit code to be added to the body bb.

* fmt

* fix same bug for for loops

* set module name to fpath