fix(diagnostics): prevent negative-length-span on empty block
Originally, the span code would for whatever reason subtract 1 from the
end of the span. This probably had a reason but I'm not sure -- I may
have just introduced a bug, but it's display code so it's not super
important. If I did, it'll be fixed by #134 if we switch to
codespan-reporting.

The diagnostic also highlighted the body of a function oddly:
```
fn f() { ... }
//       ^^^ span
```

This commit changes the function body to be represented by:
```
fn f() { ... }
//     ^^^^^^^
```

Hence, double-fixing this bug.

closes #137